Move setEnvironment to new file
